Use an access key that is known to both the server and client to sign the scores post requests.
A access key stored on the server is retrieved from the database identified by the appId in the JWT token. Using an access key that is a shared secret between client and server gives the server the ability to revoke privileges to the client by changing the accessKey on the application record.